requirements.txt
bottle==0.12.17
requests==2.22.0
python-dotenv==0.10.3
flake8>=3.7.8
flake8-docstrings>=1.4.0
pep257>=0.7.0